suite/run.py: suite_hash should use backtracked sha1
In schedule_suite() we only try to find backtracked sha1
for ceph sha1 when --newest is provided and packages are
not found in shaman. However, what about suite_sha1?
Currently, we do not use the backtracked sha1 for suite_sha1.
This commit will make sure that suite_sha1 is using
the backtracked sha1.

Fixes: https://tracker.ceph.com/issues/71612

Signed-off-by: Kamoltat Sirivadhna <ksirivad@redhat.com>